refactor: thumbnail file 관련 클래스 thumbnail/file 안으로 이동
