+++ src/main/java/com/dinosaur/foodbowl/global/util/thumbnail/file/ThumbnailFileConstants.java
@@ -1,10 +1,10 @@
-package com.dinosaur.foodbowl.global.util.thumbnail;
+package com.dinosaur.foodbowl.global.util.thumbnail.file;
 
 import static java.io.File.separator;
 
 import org.springframework.core.io.ClassPathResource;
 
-class ThumbnailConstants {
+class ThumbnailFileConstants {
 
   public static final String ROOT_PATH = "static";
   public static final String RESOURCE_PATH = new ClassPathResource(ROOT_PATH).getPath() + separator;
