+++ src/test/java/com/dinosaur/foodbowl/global/util/thumbnail/file/ThumbnailFileDtoTest.java
@@ -1,15 +1,16 @@
-package com.dinosaur.foodbowl.global.util.thumbnail;
+package com.dinosaur.foodbowl.global.util.thumbnail.file;
 
 import static com.dinosaur.foodbowl.domain.thumbnail.entity.Thumbnail.MAX_PATH_LENGTH;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.dinosaur.foodbowl.global.util.thumbnail.file.ThumbnailFileDto;
 import java.io.FileInputStream;
 import java.io.IOException;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockMultipartFile;
 
-class ThumbnailInfoDtoTest {
+class ThumbnailFileDtoTest {
 
   @Nested
   class from {
@@ -19,7 +20,7 @@ class ThumbnailInfoDtoTest {
       final MockMultipartFile tooLongFileNameMultipartFile = new MockMultipartFile("image",
           "a".repeat(MAX_PATH_LENGTH) + ".png", "image/png",
           new FileInputStream("src/test/resources/images/testImage_210x210.png"));
-      assertThatThrownBy(() -> ThumbnailInfoDto.from(tooLongFileNameMultipartFile))
+      assertThatThrownBy(() -> ThumbnailFileDto.from(tooLongFileNameMultipartFile))
           .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -28,7 +29,7 @@ class ThumbnailInfoDtoTest {
       final MockMultipartFile fakeImageFile = new MockMultipartFile("image",
           "fakeImage.png", "image/png",
           new FileInputStream("src/test/resources/images/fakeImage.png"));
-      assertThatThrownBy(() -> ThumbnailInfoDto.from(fakeImageFile))
+      assertThatThrownBy(() -> ThumbnailFileDto.from(fakeImageFile))
           .isInstanceOf(IllegalArgumentException.class);
     }
   }
